requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Config
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Auth
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
cryptography==42.0.2

# HTTP client (API aérea)
httpx==0.26.0

# Rate limiting
slowapi==0.1.9

# Processamento de arquivos (mantido)
pikepdf==8.15.1
PyMuPDF==1.24.0
moviepy==1.0.3
openai-whisper==20250625
cairosvg==2.7.1

# Testes
pytest==7.4.0
pytest-asyncio==0.23.3
